Added a series of moose test cases with different dims and element orders
