+++ scripts/gmsh/gmsh_tens_spline_2d-mod_vars.geo
