+++ scripts/gmsh/gmsh_tens_spline_3d.geo
