+++ scripts/moose-test-cases/moose-mech-subdom-3d-o1.i
@@ -1,0 +1,103 @@
+#
+# Added subdomains and subdomain-specific properties
+# https://mooseframework.inl.gov/modules/tensor_mechanics/tutorials/introduction/step03.html
+#
+
+[GlobalParams]
+  displacements = 'disp_x disp_y disp_z'
+[]
+
+[Mesh]
+  [generated]
+    type = GeneratedMeshGenerator
+    dim = 3
+    nx = 20
+    ny = 10
+    nz = 5
+    xmax = 2
+    ymax = 1
+    zmax = 1
+    elem_type = HEX8
+  []
+
+  # assign two subdomains
+  [block1]
+    type = SubdomainBoundingBoxGenerator
+    input = generated
+    block_id = 1
+    bottom_left = '0 0 0'
+    top_right = '1 1 1'
+  []
+  [block2]
+    type = SubdomainBoundingBoxGenerator
+    input = block1
+    block_id = 2
+    bottom_left = '1 0 0'
+    top_right = '2 1 1'
+  []
+[]
+
+[Modules/TensorMechanics/Master]
+  [all]
+    add_variables = true
+    generate_output = 'vonmises_stress strain_xx strain_yy strain_zz'
+  []
+[]
+
+[BCs]
+  [bottom_x]
+    type = DirichletBC
+    variable = disp_x
+    boundary = bottom
+    value = 0
+  []
+  [bottom_y]
+    type = DirichletBC
+    variable = disp_y
+    boundary = bottom
+    value = 0
+  []
+  [bottom_z]
+    type = DirichletBC
+    variable = disp_z
+    boundary = bottom
+    value = 0
+  []
+  [Pressure]
+    [top]
+      boundary = top
+      function = 1e7*t
+    []
+  []
+[]
+
+[Materials]
+  [elasticity1]
+    type = ComputeIsotropicElasticityTensor
+    youngs_modulus = 1e9
+    poissons_ratio = 0.3
+    block = 1
+  []
+  [elasticity2]
+    type = ComputeIsotropicElasticityTensor
+    youngs_modulus = 5e8
+    poissons_ratio = 0.3
+    block = 2
+  []
+  [stress]
+    type = ComputeLinearElasticStress
+  []
+[]
+
+[Executioner]
+  type = Transient
+  solve_type = 'PJFNK'
+  petsc_options_iname = '-pc_type -pc_hypre_type'
+  petsc_options_value = 'hypre boomeramg'
+  end_time = 5
+  dt = 1
+[]
+
+[Outputs]
+  exodus = true
+[]
